backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
openpyxl==3.1.2
colorama==0.4.6
jinja2==3.1.2
pywin32==306
schedule==1.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
